CMS-15008 [Backport v15] Add ApplicationSettings DaemonModule for configuring CSP settings
I've added a new DaemonModule that contains Application settings. It
currently contains the ContentSecurityPolicy settings.

To allow clients of the ApplicationSettings to be notified when the
configuration changes, the service is shutdown and re-initialized when
this happens.

(cherry picked from commit 83515881e5c2ad5965e7b430e37e4d1e9cf7c121)