CMS-15000 Allow referring documents to be selected
Java 9 Fixed a bug that skipped checking casting. See[1].

The provider is the ReferringDocumentProvider or UnpublishedReferenceNodeProvider.

The ReferringDocumentProvider provides element that implement
JcrNodeModel and the UnpublishedReferenceNodeProvider provides elements
that implements Node.

ReferringDocumentProvider#model expects a JcrNodeModel as argument and
the UnpublishedReferenceNodeProvider expects a Node.
The iterator had the type Iterator<? extends Node>, so it #next returned
an object with type Node. The implicit cast was not checked in java 8.
From java 9 onwards it does, so it throws a ClasscastException

Use a generic Iterator and provider.
UnpublishedReferenceNodeProvider#model and
ReferringDocumentProvider#model cast or wrap to always return a
Imodel<Node>.

[1] https://bugs.openjdk.java.net/browse/JDK-8058199