CMS-14960 (en|dis) (de)publication icon
The color of the circle area of the publication and depublication menu
items did not change when the "color" attribute of the enclosing div
changed, because:
- The publication icon uses the check-circle-(s|m|l|xl).svg icon. The
  fill color of the circle is a constant. (1)
- The svg "color" attribute is defined by icon-enabled and icon-disabled
  classes. (2)

Modify the (check|minus)-circle(s|m|l|xl)-clock.svg icons so that the
color of the circle area is the currentColor. It uses the "color"
attribute of the svg.

The (check|minus)-circle(s|m|l|xl)-clock.svg icon serve as inspiration.