CMS-14973 Use user instead of system session
Use user instead instead of system session, because using the user
session to save and periodically save the draft and the system session
to add, re-order and remove did give issues.

Execute the same code on the preview session as on the user session, so
the code is better maintainable.

Only save the session on the user session,  because the preview session
is thrown away after each request.