CMS-15001 Lower log level for ClientAbortionEx
Lower the log level in case the client aborts the request, so that does
not result in warning in the logs.
The ClientAbortionException is a tomcat specific exception, so we should
not add a dependency on that specific servlet specification
implemention.  That's why we check on the name of the exception.  All
other exceptions should continue to be logged as warning in the logs.

(cherry picked from commit 9c66bbc7fdea3377b4aa0b2f7a10454431d378b9)