CMS-15001 Lower log level for ClientAbortionEx
Lower the log level in case the client aborts the request, so that does
not result in warning in the logs.
The ClientAbortionException is a tomcat specific exception, so we should
not add a dependency on that specific servlet specification
implemention.  That's why we check on the name of the exception.  All
other exceptions should continue to be logged as warning in the logs.

(cherry picked from commit 2fec8a5c68a96ad3d7b5b5502ca4c79d4018e238)