CMS-15067 Support optionally to store groovy logs in repository
[backport 15.1]

By default, the reason why ENT_4612 was implemented, we won't support
writing logs to the repository when not in development mode. However,
a customer must be able to override this behavior. Besides that, when
support for writing logs to the repository is switched on, by default,
a groovy script will still write the logs to log files. However per
script, there will be the option to write the logs to the repository
by choosing Repository for the Log Target dropdown.

During development mode, writing logs to the repository instead of log
files is supported by default. For a deployment, to support this, the
system property

groovy.persist.logs.supported = true

must be set.

When this system property is present (or local) development mode, a
Groovy Script will have an extra configuration option, the Log Target.
By default, the log target will be Log Files but it can be switched to
Repository. The selected value is stored with the groovy script.

Since scripts already in history should still have their repository logs
accessible if present, showing the logs is not based on the presence of
the system property groovy.persist.logs.supported but rather on whether
there are repository stored logs available for the script or not

(cherry picked from commit f6e57e72dd6b68cf644cea501d5497fa60a7b38b)
(cherry picked from commit 64c486cf10d514419c7cc86e877221412b305529)