CMS-15077 Fix HST Model invalidation / loading
[Backport 15.1]

Next to the Asynchronous HstConfigurationEventListener add a Synchronous
listener for changes below /hst:hst. This Synchronous listener does not
synchronize on the HstModel loading of the VirtualHostsService as this
resulted in live/deadlocks. Instead it only increments an atomic long
value to keep track of the number of onEvent invocations.

When the VirtualHostsService buildNumber which is now stored does not
match the Synchronous onEventCounter while the VirtualHostsService
object in HstModelImpl has not yet been set to null, aka not yet
invalidated, it means we have to wait until the asynchronous
HstConfigurationEventListener has caught up. We cannot do this in a
method synchronized on HstModelImpl instance as this would easily result
in deadlocks. Hence the added InvalidationMonitor#awaitEventsConsistency
which waits until the async listener is at least as far as the
synchronous event listener. Only after that happens, we know all the
events for invalidating parts of the HstNode tree and Hst Configuration
objects have been processed by the HstConfigurationEventListener after
which we can continue with building the model.

Of course during building the model from the HstNode tree, the backing
JCR Nodes can already change again, making the HstModel UUIDs out of
sync with the actual JCR Nodes (in case of replacing nodes modifying
the UUIDs). However this is nothing different than always can happen
when a CMS user looks at a page in the CM : the VirtualHostsService
used to render the page can always be stale and outdated, a fact which
is just part of the interaction design and which the FE should handle
gracefully.

(cherry picked from commit 415e57e9f993531100650eac08f9d3841512488a)