CMS-15069 Fix explicit reference names support
[Backport 14.7]

This code contains a fix of a wrong functional change in CMS_14143 and
it contains an improvement in for the code:

First of all, before CMS_14143, it was supported that for pages the 'root
component' instances, the 'referencename' could be the same, so eg for
both the pages /home and /contact backed by a different hst root
component, the 'referencename' of the root component was allowed to be
for example 'base'. For child components of the root component, it is
not allowed to have duplicates as it is not allowed to have for one
single 'page instance' two different components having the same
namespace. CMS_14143, unintentionally, added this constraint to root
components as well. This commit reverts that unintended change

Next to that there was a code improvement: as we did not register which
child namespaces were in use but only had an incrementing counter, the
child components had unique autocreated names, eg r1, r2, r3, etc.
However, the code was unclear. Now instead of having a counter instance,
we properly keep track of which child namespaces are used, and create
new ones based on that

(cherry picked from commit 7d3ae501af5e2632af6978c6f030409349cdc33b)