CMS-15053 Fix SiteMapResourceTest test setup
[Backport 15.0]

The problem was very delicate, and did depend on which test ran before
the failing one. The failing test relies on a search on content based
on a hierarchical query on hippo:paths. hippo:paths is a derived
property, which however is *NOT* reset when a node is moved from a
backup location back to the original location. And this is what tests
did: in teardown, tests did move a backed expPage doc from /bak to its
original location. However, then the hippo:paths are not updated, and
searches won't find the document. Hence instead of 'move' to restore,
we need to use 'copy'.

(cherry picked from commit bc956917590046f239090f74d1cab38a3d60c440)