CMS-15080 Populate all fields used in channel grid with type equal to string
Explicitly set the data type for automatic conversion from received data to the stored value.
During sorting in the grid the sort function is selected based on the defined type.
If type equal to `string` the sorting function will be used as `asUCString` to make a case insensitive sorting.
By default if type is absent it will be defined as `auto`, see Ext.data.Field.