CMS-9587 [Backport 15.1] Refactor DialogManager configuration setup
The DialogManager maintains are reference to a default configuration
object, the DialogConfig, which it uses to create an
IPluginConfig object that is merged with an optional configuration
which is read from the request-parameters. This is to allow the
frontend to pass customized dialog options, where the manager
maintains the default options.

For CMS-9587, the dialog config can be customized based on the
context of the dialog; a second JSON string that is read from
the request-parameters and POSTed by the frontend together with the
dialogConfig.

The DialogManager will parse both JSON objects, and merge them into
an IPluginConfig. Then, it will call a new factory method
createPluginConfig(IPluginConfig configTemplate, IPluginConfig context)
to allow the user of the DialogManager to customize how the actual
dialog-config is created. By default, the configTemplate is returned.

(cherry picked from commit e5e5db83043071073ca4fa05c77c7df34588bc43)