CMS-15084 Respect mode==view in the side editor
The namespace for a document type has a "editor:templates/_root_" node.
That node has childnodes of the type "frontend:plugin" that contain
configuration options for the wicket plugin that displays the field in
the content editor. One of these properties is "mode".

When the "mode" property is "view", the field should be read-only in the
side editor.

Respect that that property for all fields for the response of the rest
call to get the content type. All field types are supported at that
level.

The UI of the site editor however for now only supports string, boolean,
long and double. (See the comments in the issue for motivation ).

(cherry picked from commit d16d067c81064317e6ddec37e494866169795510)

Improve maintainability

(cherry picked from commit 2d0aa573693855be5ed3320eacf330329d19a938)