CMS-15090 Disable variant section in the right-side panel on a foreign page
In the previous logic, when the user clicks the save button, you have to pass the component id to get the component properties from the page so that you can send the last modified date to the BE to save changes. This logic doesn't work on a different page than a selected component because it can't get the component properties from the page anymore. To fix that I pass component(which keeps all selected properties in it) to updateVariant function instead of component id.
Also, we decided to disable variant selection to avoid complexity at this point.