CMS-14628 Allow case insensitive usernames to decide to lock a container
When a user logs in to cms, a credentials object is created with the
typed
username. If the user logs in as 'foobar', then the credentials object
is
created with the username 'foobar'. By using ldap integration, a user
can
login with case insensitive usernames. When a user logs in as 'foobar'
and
locks a container, then the same user cannot edit the container again if
the user logs in again as 'FOOBAR' since a case sensitive username check
is made on hst to decide whether or not the user can edit a container.

To resolve this, the case sensitive username check to decide to lock a
container will be done by ignoring username case sensitivity.