CMS-9587 [Backport 15.1] Use $state based context in picker-service
When opening a Wicket picker from the ChannelManager, we use
the current 'params' in the $state service, to fill the dialog-context.

In most of our cases, this already provides a lot of context (channelId,
documentId, componentId, branchId). For more specific context
(like the path to a field) the PickerService can be called with a
'local' context, which will be merged with the $state context.

Note: $state params with an empty value will be skipped.
(cherry picked from commit 4dd9af745e378273811c9ea3541d9c05e29e046d)