CMS-9587 [Backport 15.1] Introduce NodePickerPluginConfig
The NodePickerPluginConfig is an IPluginConfig that is meant to be
used in the NodePickerController and has the feature that it will
interpolate the values of the input config, with values that are
available in the context,
e.g. "last.visited.key" = "${branchId}-${fieldId}".

The values in the context are best efforts, so for example, if the
context only contains the ID of a field node, we try to lookup the
document handle from that, and so on.

If a value is not available in the context, it will be replaced by an
empty string. Currently, we try to make the following variables
available (with default values).

- branchId      - The branch ID or BranchConstants.MASTER_BRANCH_ID
- channelId     - The current channel ID or empty
- componentId   - The current component ID or empty
- documentId    - The current document handle ID or empty
- documentName  - The document name or empty
- documentPath  - The absolute path to the current document or empty
- documentType  - The document type or empty
- fieldId       - The node ID of the field or empty
- fieldIndex    - The index of the field or empty
- fieldPath     - The absolute path to the field or empty
- folderPath    - The absolute path to the document folder or empty
- locale        - The user locale or Locale.ENGLISH

(cherry picked from commit 46bb6919fb8d39698141aefe8cd38e583bd01acf)