CMS-15099 Serve custom perspectives in repo order
Sync the order of custom perspectives with the order of the
configuration nodes under
/hippo:configuration/hippo:frontend/cms/cms-static, so that the ordering
complies to the documentation, see [1].

There is no constraint that a plugin that acts as perspective should
be bootstrapped under cms-static. The channel manager plugin is an example. A
client might even change the path of the cms-static node. However, if a client
wants ordered custom perspectives, the plugins have to be present in cms-static
as described in the documentation.

[1]https://documentation.bloomreach.com/14/library/concepts/editor-interface/cms-perspectives.html.

(cherry picked from commit 0489817a9acda6a02888c59b36f3672001291bc3)
(cherry picked from commit 42d0f40dbebd333fd6e5ad11c91d5af2a26864f5)