CMS-15099 Serve custom perspectives in repo order
Sync the order of custom perspectives with the order of the
configuration nodes under
/hippo:configuration/hippo:frontend/cms/cms-static, so that the ordering
complies to the documentation, see [1].

There is no constraint that a plugin that acts as perspective should
be bootstrapped under cms-static. The channel manager plugin is an example. A
client might even change the path of the cms-static node. However, if a client
wants ordered custom perspectives, the plugins have to be present in cms-static
as described in the documentation.

[1]https://documentation.bloomreach.com/14/library/concepts/editor-interface/cms-perspectives.html.

(cherry picked from commit 9df47bb6a1eed4e9e14a53b8e03cdb3966001b51)
(cherry picked from commit ae59889aba9c278aa189028edf6d908635c96213)