CMS-15122 Correct outdated logic for setting 'rendering mount id'
the logic was still based on 'FTL' rendered websites in the CM preview
which always would have an http session. However for PageDeliveryAPI (PDA)
requests, this is not the case as these requests work with a token
instead of an http session. Also in SaaS, the mounts for PDA are not
explicit as auto-created.

This code adjusts the logic that also for PDA requests with a token
instead of http session, the rendering mount id for CM preview is
set correctly