CMS-15148 Backporting ENT-5708 Introduce a cache in order to enhance performance
The getDomains() operation proved to perform extremly bad when there
are many domains and many threads attempting to perform that operation.

For that reason we are implementing a cache that will keep the result
of the operation associated to the user that the result is associated to.

Also, since the facet rules for jcr paths get resolved to uuids,
there might be a problem caused by removal and recreation of nodes,
so we have to store the path on our cache, and calculate the uuid on
runtime.