CMS-15148 Backporting ENT-5708 Make the event listener synchronous
This listener clears the cache when some changes are produced in the
jcr nodes related to the configuration of the domains.
If the listener is executed asynchronously, and the method
getFacetAuthDomains() is invoked before the listener kicks in, we will
return outdated data.
This is not so problematic on normal execution of our code, however,
it causes issues when running tests, because sometimes the expectations
are not met.

Using synchronous JCR event listeners can be problematic when performing
more JCR operations as part of the listener, however in this case we are
just invalidating the cache, so it should be safe.