CMS-15144 Never redirect to a host which is not configured
a redirect which does not match in the configured host
configuration below hst:platform can never be a valid host. Therefore,
we do check the farthest request host (X-Forwarded-Host header) whether
the provided host can be matched in the hst:platform hosts, and if not,
we return a 400 as it is definitely an invalid host.

Also this issue fixes that in case the cmsSessionContext is null, a 400
is returned instead of a 200 and a long stacktrace being logged by
SwallowExceptionMapper