CMS-15203 Implement an on request filtered cached SiteMap tree approach
[Backport 15.2]

Initial note : The implementation with multiple different POJO objects
are mostly a result from legacy code and backward compatible support for
the old style to keep working. There are many intermediate POJOs which
might seem unhandy, and which also are unhandy, but this is the reason
why they are still there.

Functional:

Next to the 'resty' endpoint

http://localhost:8080/cms/_rp/{id}./pagetree

we now support

http://localhost:8080/cms/_rp/{id}./sitemapitem

which returns the 'root' sitemap item plus its direct children in which
the children now also have a flag 'expandable' to indicate whether they
can be expanded

For deeper nested SiteMap items, there are resty endpoints for the
pathInfo elements of SiteMap Items. For example

http://localhost:8080/cms/_rp/{id}./sitemapitem/products

returns the SiteMap item 'products' plus its direct children.

Implementation global:

- A general 'read-everywhere' cached in memory model for all XPage docs
is added
- This read-everywhere cached model is per request filtered for a CMS
User : the filtering must be done because the CMS user might not have
read-access to all XPage Documents. As we only return the direct
children below some SiteMap item (instead of the entire tree), we need
to figure out whether a SiteMap item is expandable. This is done by
checking whether the user has at least one readable descendant for the
SiteMap item: if so, expandable = true.
- After the filter XPage SiteMap Shallow (only direct children) is
retrieved, the HST sitemap (routes) are glued in if not yet present

Implementation details:

- A general 'read-everywhere' cached in memory model for all XPage docs
of a channel is implemented in XPageSiteMapRepresentationService. It
reads all XPages for a channel (branch) once, and then constructs a URL
space (sitemap tree) for it from the individual pathInfo of every XPage
doc.

Two invalidation mechanisms are in place, which trigger for now a full
reload of the XPage SiteMap tree for a channel. In the future, we might
be chose for a more efficient fine-grained reloading but this is quite
some delicate work.

- Invalidation algo 1 : The HST sitemap (routes) of a channel impacts
the URLs for XPage docs. When the sitemap changes, it means the
resulting XPage SiteMap tree also typically changes for XPage docs as
the linkrewriting for an XPage might change as a result. For that reason,
on HST sitemap (routes) change, the XPage SiteMap tree most be reloaded.
The HST sitemap (routes) can be different between CORE channel and a
branch, hence also the resulting XPage SiteMap tree. What makes it even
more complex, is that HST sitemap (routes) also inherit from eg the
hst:default/hst:sitemap. When something here changes, also the SiteMap
Tree most be reloaded. To support this,

Object o = HstConfigurationLoadingCache#getSiteMapConfigurationIdentity

has been introduced which returns an Object instance which keeps being
the same as long as the HST sitemap (routes) for a channel (branch)
did not change: this piggybacks on the HST model invalidation logic
for HstNode. The returned object 'o' represents the 'cache key' for the
SiteMap Tree. This is used in XPageSiteMapRepresentationService#cache
as key. Whenever the HST sitemap (routes), the object identity returned
from HstConfigurationLoadingCache#getSiteMapConfigurationIdentity changes,
triggering a cache miss. The stale entry in the cache will be cleared
later through XPageSiteMapRepresentationService#expungeStaleEntries

- Invalidation algo 2: A JCR listener for node ADD or node REMOVE below
hippostd:folder has been added. Listening to MOVE is not needed as this
is always accompanied with a ADD and REMOVE event. The events are
collected and when a new XPAge SiteMap Tree instance is requested, the
events are processed. When an XPage has been added or removed, the previous
XPAge SiteMap Tree is considered stale and most be updated/reloaded.

(cherry picked from commit f051b3b9043f51b383220341feced8770bf13f57)