CMS-15204 In case item for pathInfo is not found, do not expand
[Backport 15.2]

Do not expand to farthest found ancestor in case the item for pathInfo
cannot be found. Note that instead of a 404 a 400 is returned. This is
something we should address separately as every client exception is
currently returned as a 400 instead of more precise error code

(cherry picked from commit 931f0db4e561af9f13a6bfeeeb88efb0cc4b316c)