ENT-4860 Support targeted component rendering URL for XPage documents
[Forward port 15.2]

Component rendering URL of an XPage did not pick a component variant in
the Delivery API, while the full URL does it correctly. The former
always did a fallback to the default component variant.

The problem turned out to be that

Collection<Persona> allPersonasForChannel
___________________ = getPersonas(componentConfiguration.getMountVariants(), config);

in com.onehippo.cms7.targeting.TargetingServiceImpl#createTargetingProfile
returned 0 available Persona objects for component rendering URLs for an
XPage document. This was the result that in ExperiencePageServiceImpl only
to the 'root page component' the available MountVariants were updated with

copy.getMountVariants().addAll(copy.getVariants());

Instead of the above, now the code invokes

copy.addMountVariants(copy.getVariants());

and the 'addMountVariants' makes sure to traverse also all the child
components if the components is either detached OR is an
ExperiencePageComponent

(cherry picked from commit fdb48f89656dd1273ef7b321cce8ac21d28d8695)