CMS-15184 Add package-info to the component-beans node_modules
The problem seems somehow related to the changes on the JVM after
version 8.

The reason why this problem occurs is explained in
 https://docs.oracle.com/javase/9/docs/api/java/lang/Package.html#getPackage-java.lang.String-

> If multiple class loaders delegate to each other and define classes
with the same package name, and one such loader relies on the lookup
behavior of getPackage to return a Package from a parent loader, then
the properties exposed by the Package may not be as expected in the
rest of the program. For example, the Package will only expose
annotations from the package-info.class file defined by the parent
loader, even if annotations exist in a package-info.class file defined
by a child loader.

Our content beans inherit from
- org.hippoecm.hst.content.beans.standard.HippoDocument
- org.hippoecm.hst.content.beans.standard.HippoItem
- org.hippoecm.hst.content.beans.standard.HippoBean

Note that the same package is used, but HippoDocument & HippoItem are
part of content-beans module (which is deployed on the site), while
HippoBean is part of the hst-api (deployed on the shared classloader)

We already had a package-info class specifying not to serialize
anything in the package org.hippoecm.hst.content.beans.standard by
default, so the properties of the aforementioned classes are not
serialized.

That package-info is defined in the module hst-api, which is loaded by
the shared classloader. Before java 9, the JVM was able to load the
proper package info, from the shared classloader, even when working
with HippoDocument & HippoItem (which are loaded by the site).

Since java 9, the JVM is not able to load the necessary information, so
it tries to serialize every field of HippoDocument and HippoItem,
which includes entities such as jcr Node or ObjectConverter that
shouldn't be serialized. This is what introduced the error that is
being fixed in this commit.

There is no easy solution here, but an easy and effective workaround is
to duplicate the package-info, including it in the artifact
content-beans (where HippoDocument & HippoItem are defined), so it gets
loaded by the site classloader.