CMS-15199 Allow managing external preview/urls on channel settings page
External preview can be enabled over the channel settings page. Also the
token can be regenerated on the same page. This can be managed on core
and the projects separately. Also a preview button is added to the
toolbar on the channel manager view.

The site application can also be a ftl view site so besides spa sites,
this type of site is also supported while preparing the preview urls
which is different than SaaS.

There is another difference on spa url configuration between PaaS and
SaaS. In PaaS, spa url may also contain the endpoint url parameter which
is not expected in SaaS. This distinction is also handled properly.