CMS-15199 Add a new sub menu to page menu to copy page preview url
On channel manager view, a new sub menu item is added to the ‘Page’ menu
to be able copy page preview url to the clipboard.
This menu item is available when the channel preview is enabled for the
core or a specific project.

(cherry picked from commit fb1bae6a44c0ded4adb4f6bdade378a5d5cbabc6)