CMS-15127 For XPage deep copy, inherit 'compositeCacheable' as well
[Backport 14.7]

By default, compositeCacheable instance variable is true and during
the hst model build, it is deducted from all descendant components
whether the 'root page component' is compositeCacheable or not : this is
the case when all descendants and the component itself are cacheable
(async component skipped).

However, during a 'deepCopy' for an XPage document component, the
compositeCacheable value was not inherited, which caused the XPage doc
to be compositeCacheable even if it got cloned from an uncacheable
layout component. This is fixed by also inheriting the
compositeCacheable

(cherry picked from commit c4d4599993dc1a5537de9dd11fea44a0ce05f97b)