CMS-15196 Show components overlay by default
Set the default state to true for components, so that
- the state will be set to true in
  OverlayToggleController.initiateOverlay if:
  - the value of the "isInitiallyDisplayed" binding is false; and
  - the "ls.channelManager.overlays.components" parameter is not yet
    present in the local storage.
so that if a user logs in for the first time the component overlay is
shown if "isInitiallyDisabled" is false ( which in almost all situations
is the case ). See ChannelCtrl.isComponentOverlayInitiallyDisabled.

The component overlay now has almost the same behavior as the content
overlay toggle. The only difference is that the component overlay can
also be enable is the project is in review and "resetChannel" is allowed
for the current user. See ProjectService.isEditingAllowed.

(cherry picked from commit fb57720be82138266ee1c6f7cf24112dbf1cd4c4)