CMS-15213 Support pluggable groovy CompilationCustomizers
An end project can register a service in the HippoServiceRegistry which
implements CompilationCustomizerFactory. When this service is available,
the service is used instead of the default used CompilationCustomizers
from GroovyUpdaterClassLoader.

To support easy / reuse for end projects, a couple of static class
variables and methods have been made public. To support 'Logger' without
importing it also 'org.slf4j' has been added to the default star imports