CMS-15343 Fix privilege check for XPage Docs in Channel SiteMap
the userrole xm.content.viewer is by default used to assign jcr:read
privilege to specific (content) jcr nodes. However, at this moment, the
privilege we currently check on whether the user is allowed to see an
XPage doc in the channel SiteMap is hippo:author. The reason is that
we did not anticipate customers using the 'xm.content.viewer' userrole
ever but always either xm.content.author or xm.content.editor or
xm.content.admin. As the only real relevant check is read access, we can
simply change the privilege check from hippo:author to jcr:read