CMS-15312 Only log warning when events are still out of sync after 60 sec
[Backport 15.2]

In case the async and sync listener are not in sync within 1 second, a
warning was logged. However, we cannot know whether this is a persistent
missed async event or a temporal missing event.

For that reason, we for now log an info and only start logging a warn
in case the problem persists longer than 60 seconds

(cherry picked from commit 5646f52f0feb1b3d2573f5d91147e870e21afefa)